refactor: make PrintableEventReport into server component
- Move the styled jsx into globals
- Move Event and Student types to global type files
- Cleanup imports
